Separate GitRepoHelper into local and remote classes
